Removed unneeded boxing/unboxing and some other redundancies. Other minor language improvements.

+++ app/src/main/java/org/xbill/DNS/ResolverConfig.java
@@ -372,7 +372,7 @@ private void
 findWin(InputStream in) {
 	String property = "org.xbill.DNS.windows.parse.buffer";
 	final int defaultBufSize = 8 * 1024;
-	int bufSize = Integer.getInteger(property, defaultBufSize).intValue();
+	int bufSize = Integer.getInteger(property, defaultBufSize);
 	BufferedInputStream b = new BufferedInputStream(in, bufSize);
 	b.mark(bufSize);
 	findWin(b, null);
